requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2
django-admin-sortable2==2.2.3
django-modeltranslation==0.19.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
package_name==0.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
